+++ src/lib/containers/__stories__/index.ts
@@ -1,2 +1,3 @@
 import "./artist.story"
 import "./gene.story"
+import "./my_account.story.tsx"