+++ src/lib/containers/__stories__/my_account.story.tsx
@@ -1,0 +1,12 @@
+import { storiesOf } from "@kadira/react-native-storybook"
+import * as React from "react"
+import { RootContainer } from "react-relay"
+
+import Routes from "../../relay/routes"
+import MyAccount from "../my_account"
+
+storiesOf("My Account")
+  .add("Root", () => {
+    const profileRoute = new Routes.MyAccount()
+    return <RootContainer Component={MyAccount} route={profileRoute} />
+  })